\name{cpptf}
\alias{cpptf}
\docType{package}
\title{\packageTitle{cpptf}}
\description{\packageDescription{cpptf}}
\details{The DESCRIPTION file: \packageDESCRIPTION{cpptf}}
\section{Package Content}{\packageIndices{cpptf}}
\author{\packageAuthor{cpptf}}
\section{Maintainer}{\packageMaintainer{cpptf}}
\references{https://github.com/cpp-taskflow/cpp-taskflow}